validation 추가, websocket 수정

+++ arom1/src/main/java/com/example/arom1/controller/MeetingController.java
@@ -5,22 +5,18 @@ import com.example.arom1.common.exception.BaseException;
 import com.example.arom1.common.response.BaseResponse;
 import com.example.arom1.common.response.BaseResponseStatus;
 import com.example.arom1.dto.MeetingDto;
-import com.example.arom1.dto.ReviewDto;
 import com.example.arom1.dto.response.MeetingResponse;
-import com.example.arom1.dto.response.ReviewResponse;
-import com.example.arom1.entity.Meeting;
-import com.example.arom1.entity.MemberDetail;
-import com.example.arom1.entity.Review;
 import com.example.arom1.repository.MeetingRepository;
 import com.example.arom1.service.MeetingService;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
-import java.net.URLEncoder;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/eatery")
@@ -28,8 +24,7 @@ public class MeetingController {
 
     @Autowired
     private MeetingService meetingService;
-    @Autowired
-    private MeetingRepository meetingRepository;
+
 
     //음식점 선택 후 식사 메이트 구하는 글 보기
     @GetMapping("/{eateryId}/meetings")
@@ -49,7 +44,12 @@ public class MeetingController {
 
     //음식점 식사 메이트 구하는 글 작성
     @PostMapping("/{eateryId}/meetings")
-    public BaseResponse<?> saveMeeting(@PathVariable long eateryId, @RequestBody MeetingDto meetingDto){
+    public BaseResponse<?> saveMeeting(@PathVariable long eateryId, @Valid @RequestBody MeetingDto meetingDto, BindingResult bindingResult){
+       if(bindingResult.hasErrors()){
+           List<FieldError> errors = bindingResult.getFieldErrors();
+           return new BaseResponse<>(HttpStatus.BAD_REQUEST.value());
+       }
+
         try{
             // 저장된 회의 정보를 반환
             MeetingResponse meetingResponse = meetingService.saveMeeting(meetingDto);
@@ -67,7 +67,11 @@ public class MeetingController {
 
     //수정
     @PutMapping("/{eateryId}/meetings/{meetingId}")
-    public BaseResponse<?> updateMeeting(@PathVariable Long eateryId, @PathVariable Long meetingId, @RequestBody MeetingDto meetingDto ){
+    public BaseResponse<?> updateMeeting(@PathVariable Long eateryId, @PathVariable Long meetingId, @Valid @RequestBody MeetingDto meetingDto, BindingResult bindingResult ){
+        if(bindingResult.hasErrors()){
+            List<FieldError> errors = bindingResult.getFieldErrors();
+            return new BaseResponse<>(HttpStatus.BAD_REQUEST.value());
+        }
         try{
             MeetingResponse updatedMeeting = meetingService.updateMeeting(eateryId, meetingId, meetingDto);
             return new BaseResponse<>(updatedMeeting);
@@ -79,7 +83,11 @@ public class MeetingController {
 
     //삭제
     @DeleteMapping("/{eateryId}/meetings/{meetingId}")
-    public BaseResponse<?> deleteMeeting(@PathVariable Long eateryId, @PathVariable Long meetingId, @RequestBody MeetingDto meetingDto ){
+    public BaseResponse<?> deleteMeeting(@PathVariable Long eateryId, @Valid @PathVariable Long meetingId, @RequestBody MeetingDto meetingDto, BindingResult bindingResult ){
+        if(bindingResult.hasErrors()){
+            List<FieldError> errors = bindingResult.getFieldErrors();
+            return new BaseResponse<>(HttpStatus.BAD_REQUEST.value());
+        }
         try{
             meetingService.deleteMeeting(meetingId, meetingDto.getMember_id());
             return new BaseResponse<>(BaseResponseStatus.SUCCESS);
